Switched to buildroot-based toolchain.
The motivation is the upcoming integration of Linux. Therefore, we want
to be able to use the same toolchain as for M³. The previous self-build
toolchain was specifically build for M³ and therefore not directly
usable for Linux. Additionally, switching to buildroot has also the
advantage that we can more easily upgrade it and do no longer have to
maintain patches for binutils, gcc, and gdb.

The switch was far from straight forward, though. The primarily problem
was caused by our requirement to build TileMux with soft-float. On
RISC-V, there is both an architecture (-march) that defines the
available extensions and an ABI (-mabi). Previously, we used the ABI
lp64, which does not pass floating point values in registers and thus
is agnostic to whether the f and d extensions are available or not.
However, the buildroot toolchain and in particular the Rust toolchain
for which a Rust standard library is available uses lp64d. This ABI
passes floats in registers and thus requires the f/d extension to be
available (otherwise the registers don't exist). Obviously, it is
impossible to link object files with different ABIs, because they would
be incompatible. So, in summary, we were able to sidestep this problem
by using lp64 (since we don't really care about floats anyway), but now
we have to use lp64d and therefore need a solution.

The solution we use now is to build TileMux completely separate from
everything else. In other words, we build TileMux itself including all
dependencies (again) with soft-float. For Rust it means that we use a
separate target spec (as there seems to be no way to control the ABI
via flags to rustc) for soft-float. This conveniently also builds all
dependencies including core, alloc etc. with soft-float. For C/C++ it's
more complicated (as always). Fortunately, TileMux does not depend on
any toolchain-provided libraries (gcc, etc.). We therefore only need to
build our own libraries twice (simplec, gem5, isr).

To avoid a complete mess in the build scripts, I introduce the ability
to build the same source files with different environments without
interference (see other commit), and a method on Env that enables
soft-float.

In the end, it did not cause any major trouble and arguably we should
have already build TileMux separately anyway.

Furthermore, we are now required to provide rudimentary support for
thread-local storage. This is used by the libsupc++ during exception
handling, which was previously disabled, because we built libsupc++
with --disable-tls. Now that we want to use the standard Linux
toolchain, we do no longer want to disable it and therefore need
support. This requires some initialization in musl and, unfortunately,
on x86_64 also a new TMCall in order to set the fs segment register.

+++ src/apps/bench/cppbenchs/benchs/bsyscall.cc
@@ -50,7 +50,8 @@ NOINLINE static void activate() {
 }
 
 NOINLINE static void create_mgate() {
-    static uintptr_t addr = Math::round_dn(reinterpret_cast<uintptr_t>(&create_mgate), PAGE_SIZE);
+    static uintptr_t addr = Math::round_dn(reinterpret_cast<uintptr_t>(&create_mgate),
+                                           static_cast<uintptr_t>(PAGE_SIZE));
 
     struct SyscallMGateRunner : public Runner {
         void run() override {
